chore: merge 2 components to 1
